CryptoStat/backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
binance-connector==3.7.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
dateparser==1.2.0
Django==5.0.6
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
multidict==6.0.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyOpenSSL==24.1.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
zope.interface==6.4.post2
